First revision.  Fails to reproduce error however.
So I thought this was a minimal representation of what we're doing, but
it fails to reproduce the issue.  I guess I'll have to hook in mocha and
see if that starts causing the issue.

Alternatively, it could be a ubuntu specific issue.  I first tested this
in my base OS instead of setting up a vagrant box.