Updated case to where the errant code is run by mocha + added vagrant
Since I wasn't reproducing the issue in my initial tests, I added mocha
into the mix.  That didn't reproduce the issue either.  However, adding
a ubuntu VM via vagrant to the case with mocha DID start reproducing the
issue.  The call to child_process.exec started creating 2 processes, the
first of which was a call to sh with the -c argument to execute the
actual command, presumably done this way so as to set the working
directory as specified in my code.

I'll strip out the working directory option to see if that makes the
issue go away.